refactor: tidy initial implementation
Move prefetch and network information functions into
seperate files. Create useDelay function to remove
deeply nested if/else statements in intersection
observer callback.
