+++ lib/index.ts
@@ -1,43 +1,73 @@
-// NetworkInformation API not currently included
-// in typescript DOM types
-type EffectiveConnectionType = "slow-2g" | "2g" | "3g" | "4g";
-type ConnectionType =
-    | "bluetooth"
-    | "cellular"
-    | "ethernet"
-    | "none"
-    | "wifi"
-    | "wimax"
-    | "other"
-    | "unknown";
+import { isSaveDataEnabled, isSlowConnectionType } from "./network-information";
+import { prefetch } from "./prefetch";
 
-interface NetworkInformation extends EventTarget {
-    readonly downlink: number;
-    readonly downlinkMax: number;
-    readonly effectiveType: EffectiveConnectionType;
-    readonly rtt: number;
-    readonly saveData: boolean;
-    readonly type: ConnectionType;
+interface Options {
+    elements?: HTMLAnchorElement | NodeList;
+    selectors?: string[];
+    ignore?: string[];
+    delayMilliseconds?: number;
+    inViewThreshold?: number;
 }
 
-function isSaveDataEnabled(): boolean {
-    if ("connection" in window.navigator) {
-        const connection = window.navigator.connection as NetworkInformation;
-        return connection.saveData;
+export function watch(options?: Options): void {
+    if (isSaveDataEnabled()) {
+        return;
     }
 
-    return false;
-}
+    // sets options or uses defaults
+    const elements = options?.elements;
+    const selectors = options?.selectors;
+    const ignore = options?.ignore || [];
+    const delayMilliseconds = options?.delayMilliseconds || 0;
+    const inViewThreshold = options?.inViewThreshold || 0;
 
-function isSlowConnectionType(): boolean {
-    if ("connection" in window.navigator) {
-        const connection = window.navigator.connection as NetworkInformation;
-        return (
-            connection.effectiveType === "slow-2g" ||
-            connection.effectiveType === "2g"
-        );
-    }
-    return false;
+    const idleCallback = useIdleCallback();
+
+    const toFetch = new Set<string>();
+    const fetched = new Set<string>();
+
+    const observer = new IntersectionObserver(
+        (entries, observer) => {
+            entries.forEach((entry) => {
+                const link = entry.target as HTMLAnchorElement;
+
+                // in viewport and not yet prefetched
+                if (entry.isIntersecting && !fetched.has(link.href)) {
+                    toFetch.add(link.href);
+
+                    // uses setTimeout if delayMilliseconds > 0 or
+                    // immediately calls callback
+                    useDelay(() => {
+                        if (
+                            entry.isIntersecting &&
+                            toFetch.has(link.href) &&
+                            !isSlowConnectionType()
+                        ) {
+                            idleCallback(() => {
+                                prefetch(toFetch);
+                            });
+
+                            fetched.add(link.href);
+                            observer.unobserve(entry.target);
+                        }
+                    }, delayMilliseconds);
+
+                    // no longer in viewport
+                } else {
+                    toFetch.delete(link.href);
+                }
+            });
+        },
+        { threshold: inViewThreshold },
+    );
+
+    const links = assembleListOfLinks(elements, selectors);
+
+    links.forEach((link) => {
+        if (!shouldIgnore(new URL(link.href), ignore)) {
+            observer.observe(link);
+        }
+    });
 }
 
 function shouldIgnore(url: URL, ignoreList: string[]): boolean {
@@ -49,7 +79,7 @@ function shouldIgnore(url: URL, ignoreList: string[]): boolean {
         return true;
     }
 
-    // can't return from inside a forEach
+    // can't return from inside a forEach loop
     let shouldIgnore = false;
     ignoreList.forEach((entry) => {
         if (url.href.includes(entry)) {
@@ -58,72 +88,6 @@ function shouldIgnore(url: URL, ignoreList: string[]): boolean {
     });
 
     return shouldIgnore;
-}
-
-function supportsSpecRulesAPI(): boolean {
-    return (
-        HTMLScriptElement.supports &&
-        HTMLScriptElement.supports("speculationrules")
-    );
-}
-
-function appendSpecRuleToDOM(hrefs: Set<string>): void {
-    const specScript = document.createElement("script");
-    specScript.type = "speculationrules";
-    const specRules = {
-        prefetch: [
-            {
-                source: "list",
-                urls: [...hrefs],
-            },
-        ],
-    };
-
-    specScript.textContent = JSON.stringify(specRules);
-    document.body.append(specScript);
-}
-
-function supportsPrefetchResourceHint() {
-    const link = document.createElement("link");
-    return (
-        link.relList &&
-        link.relList.supports &&
-        link.relList.supports("prefetch")
-    );
-}
-
-function appendPrefetchResourceHintToDOM(hrefs: Set<string>): void {
-    hrefs.forEach((href) => {
-        const link = document.createElement("link");
-        link.href = href;
-        link.rel = "prefetch";
-
-        document.head.appendChild(link);
-    });
-}
-
-function prefetchUsingFetchAPI(hrefs: Set<string>) {
-    hrefs.forEach((href) => {
-        const options: RequestInit = {
-            priority: "low",
-            headers: {
-                Purpose: "prefetch",
-            },
-        };
-
-        fetch(href, options);
-    });
-}
-
-function prefetch(toFetch: Set<string>): void {
-    if (supportsSpecRulesAPI()) {
-        appendSpecRuleToDOM(toFetch);
-    } else if (supportsPrefetchResourceHint()) {
-        appendPrefetchResourceHintToDOM(toFetch);
-    } else {
-        prefetchUsingFetchAPI(toFetch);
-    }
-    toFetch.clear();
 }
 
 function isNodeAnchorElement(node: Node): boolean {
@@ -135,6 +99,7 @@ function assembleListOfLinks(
     selectors?: string[],
 ): HTMLAnchorElement[] {
     const links: HTMLAnchorElement[] = [];
+
     if (elements) {
         if ("length" in elements) {
             elements.forEach((ele) => {
@@ -142,22 +107,20 @@ function assembleListOfLinks(
                     links.push(ele as HTMLAnchorElement);
                 }
             });
-        } else {
-            if (isNodeAnchorElement(elements)) {
-                links.push(elements);
-            }
+        } else if (isNodeAnchorElement(elements)) {
+            links.push(elements);
         }
     }
-    if (selectors) {
-        selectors.forEach((selector) => {
-            const nodes = document.querySelectorAll(selector);
-            nodes.forEach((node) => {
-                if (isNodeAnchorElement(node)) {
-                    links.push(node as HTMLAnchorElement);
-                }
-            });
+
+    selectors?.forEach((selector) => {
+        const nodes = document.querySelectorAll(selector);
+        nodes.forEach((node) => {
+            if (isNodeAnchorElement(node)) {
+                links.push(node as HTMLAnchorElement);
+            }
         });
-    }
+    });
+
     if (!elements && !selectors) {
         const anchorTags = document.querySelectorAll("a");
         anchorTags.forEach((tag) => links.push(tag));
@@ -184,73 +147,11 @@ function useIdleCallback() {
     );
 }
 
-interface Options {
-    elements?: HTMLAnchorElement | NodeList;
-    selectors?: string[];
-    ignore?: string[];
-    delayMilliseconds?: number;
-    inViewThreshold?: number;
-}
-
-export function watch(options?: Options): void {
-    if (isSaveDataEnabled()) {
+function useDelay(callback: () => void, delayMilliseconds: number) {
+    if (delayMilliseconds === 0) {
+        callback();
         return;
     }
-    // sets options or uses defaults
-    const elements = options?.elements;
-    const selectors = options?.selectors;
-    const ignore = options?.ignore || [];
-    const delayMilliseconds = options?.delayMilliseconds || 0;
-    const inViewThreshold = options?.inViewThreshold || 0;
-    const idleCallback = useIdleCallback();
 
-    const toFetch = new Set<string>();
-    const fetched = new Set<string>();
-
-    const observer = new IntersectionObserver(
-        (entries, observer) => {
-            entries.forEach((entry) => {
-                const link = entry.target as HTMLAnchorElement;
-                if (entry.isIntersecting) {
-                    if (!fetched.has(link.href)) {
-                        if (delayMilliseconds > 0) {
-                            toFetch.add(link.href);
-                            setTimeout(() => {
-                                if (
-                                    entry.isIntersecting &&
-                                    toFetch.has(link.href)
-                                ) {
-                                    idleCallback(() => {
-                                        prefetch(toFetch);
-                                    });
-                                    fetched.add(link.href);
-                                    observer.unobserve(entry.target);
-                                }
-                            }, delayMilliseconds);
-                        } else {
-                            if (!isSlowConnectionType()) {
-                                toFetch.add(link.href);
-                                idleCallback(() => {
-                                    prefetch(toFetch);
-                                });
-                                fetched.add(link.href);
-                                observer.unobserve(entry.target);
-                            }
-                        }
-                    }
-                } else {
-                    toFetch.delete(link.href);
-                }
-            });
-        },
-        { threshold: inViewThreshold },
-    );
-
-    const links = assembleListOfLinks(elements, selectors);
-
-    links.forEach((link) => {
-        if (!shouldIgnore(new URL(link.href), ignore)) {
-            observer.observe(link);
-        }
-    });
+    setTimeout(callback, delayMilliseconds);
 }
